+++ src/frost/redjubjub.rs
@@ -19,7 +19,7 @@ pub use frost_rerandomized::frost_core::{
 };
 pub use rand_core;
 
-use rand_core::{CryptoRng, RngCore};
+use rand_core::{CryptoRng, Rng};
 
 use crate::{hash::HStar, private::Sealed, sapling};
 
@@ -53,7 +53,7 @@ impl Field for JubjubScalarField {
         }
     }
 
-    fn random<R: RngCore + CryptoRng>(rng: &mut R) -> Self::Scalar {
+    fn random<R: Rng + CryptoRng>(rng: &mut R) -> Self::Scalar {
         Self::Scalar::random(rng)
     }
 
@@ -219,7 +219,7 @@ pub mod keys {
 
     /// Allows all participants' keys to be generated using a central, trusted
     /// dealer.
-    pub fn generate_with_dealer<RNG: RngCore + CryptoRng>(
+    pub fn generate_with_dealer<RNG: Rng + CryptoRng>(
         max_signers: u16,
         min_signers: u16,
         identifiers: IdentifierList,
@@ -234,7 +234,7 @@ pub mod keys {
     /// instead of generating a fresh one. This is useful in scenarios where
     /// the key needs to be generated externally or must be derived from e.g. a
     /// seed phrase.
-    pub fn split<R: RngCore + CryptoRng>(
+    pub fn split<R: Rng + CryptoRng>(
         key: &SigningKey,
         max_signers: u16,
         min_signers: u16,
@@ -321,7 +321,7 @@ pub mod round1 {
         rng: &mut RNG,
     ) -> (SigningNonces, SigningCommitments)
     where
-        RNG: CryptoRng + RngCore,
+        RNG: CryptoRng + Rng,
     {
         frost::round1::commit::<J, RNG>(secret, rng)
     }