+++ tests/frost_redpallas.rs
@@ -2,10 +2,10 @@
 
 use std::collections::BTreeMap;
 
-use group::GroupEncoding;
-use rand::thread_rng;
-
 use frost_rerandomized::frost_core::{self as frost, Ciphersuite, Group, GroupError};
+use group::GroupEncoding;
+
+use rand::rng as thread_rng;
 
 use reddsa::{
     frost::redpallas::{keys::EvenY, PallasBlake2b512},